backend/requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
asyncpg==0.30.0
bcrypt==4.2.1
blinker==1.8.2
cffi==1.17.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
databases==0.9.0
debugpy==1.8.11
decorator==5.1.1
distlib==0.3.9
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.6
filelock==3.17.0
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.55.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jwt==1.3.1
kiwisolver==1.4.8
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
motor==3.7.0
neo4j==5.25.0
nest-asyncio==1.6.0
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
pywin32==308
pyzmq==26.2.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
virtualenv==20.29.1
wcwidth==0.2.13
Werkzeug==3.1.0
